move the wrapper_creator code into a dedicated module
