+++ bfabric/scripts/bfabric_setWorkunitStatus_failed.py
@@ -19,10 +19,11 @@ import bfabric
 from random import randint
 from time import sleep
 
+import bfabric.wrapper_creator.bfabric_feeder
 
 if __name__ == "__main__":
     if len(sys.argv) > 1:
-        bfapp = bfabric.BfabricFeeder()
+        bfapp = bfabric.wrapper_creator.bfabric_feeder.BfabricFeeder()
 
         workunitid = int(sys.argv[1])
         print("workunitit={}".format(workunitid))
